requirements.txt
sanic==23.12.1
sanic-ext==23.12.0
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
pyjwt==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
email-validator==2.1.0
pytest==7.4.4
pytest-asyncio==0.23.2
requests==2.31.0 
